requirements.txt
PyQt6==6.7.1
requests==2.31.0
Pillow==10.4.0
psutil==5.9.8
